+++ packages/web-components/src/components/background-media/defs.ts
@@ -1,7 +1,7 @@
 /**
  * @license
  *
- * Copyright IBM Corp. 2021
+ * Copyright IBM Corp. 2021, 2025
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -15,6 +15,11 @@ export enum GRADIENT_DIRECTION {
    * Left to right
    */
   LEFT_TO_RIGHT = 'left-to-right',
+
+  /**
+   * Right to left
+   */
+  RIGHT_TO_LEFT = 'right-to-left',
 
   /**
    * Top to Bottom