+++ packages/web-components/src/components/callout-quote/__stories__/callout-quote.stories.ts
@@ -1,7 +1,7 @@
 /**
  * @license
  *
- * Copyright IBM Corp. 2020, 2023
+ * Copyright IBM Corp. 2020, 2025
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -15,6 +15,7 @@ import { QUOTE_TYPES } from '../../quote/quote';
 import { COLOR_SCHEME } from '../../../component-mixins/callout/defs';
 import readme from './README.stories.mdx';
 import textNullable from '../../../../.storybook/knob-text-nullable';
+import { APPEARANCE } from '../../quote/defs';
 
 const quoteTypes = {
   [`${QUOTE_TYPES.DEFAULT}`]: QUOTE_TYPES.DEFAULT,
@@ -24,6 +25,12 @@ const quoteTypes = {
   [`${QUOTE_TYPES.LOW_HIGH_REVERSED_DOUBLE_CURVED}`]:
     QUOTE_TYPES.LOW_HIGH_REVERSED_DOUBLE_CURVED,
   [`${QUOTE_TYPES.CORNER_BRACKET}`]: QUOTE_TYPES.CORNER_BRACKET,
+};
+
+const appearance = {
+  [`${APPEARANCE.DEFAULT}`]: APPEARANCE.DEFAULT,
+  [`${APPEARANCE.CARD}`]: APPEARANCE.CARD,
+  [`${APPEARANCE.BUBBLE_QUOTE}`]: APPEARANCE.BUBBLE_QUOTE,
 };
 
 const colorSchemeTypes = {
@@ -38,13 +45,17 @@ export const Default = (args) => {
   const {
     copy,
     quoteMark,
+    appearance,
     sourceHeading,
     sourceCopy,
     sourceBottomCopy,
     colorScheme,
   } = args?.CalloutQuote ?? {};
   return html`
-    <c4d-callout-quote mark-type="${quoteMark}" color-scheme="${colorScheme}">
+    <c4d-callout-quote
+      mark-type="${quoteMark}"
+      appearance="${appearance}"
+      color-scheme="${colorScheme}">
       ${copy}
       <c4d-quote-source-heading> ${sourceHeading} </c4d-quote-source-heading>
       <c4d-quote-source-copy> ${sourceCopy} </c4d-quote-source-copy>
@@ -86,6 +97,11 @@ export default {
           'Quote Mark (markType):',
           quoteTypes,
           quoteTypes['double-curved']
+        ),
+        appearance: select(
+          'Appearance (card and bubble-quote - MOBILE ONLY | Default - legacy):',
+          appearance,
+          appearance['legacy']
         ),
         sourceHeading: textNullable(
           'Source heading (source-heading slot)',