+++ packages/web-components/src/components/quote/defs.ts
@@ -1,7 +1,7 @@
 /**
  * @license
  *
- * Copyright IBM Corp. 2020, 2021
+ * Copyright IBM Corp. 2020, 2025
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -37,4 +37,21 @@ export enum QUOTE_TYPES {
    * corner-bracket
    */
   CORNER_BRACKET = 'corner-bracket',
+}
+
+export enum APPEARANCE {
+  /**
+   * Default - legacy
+   */
+  DEFAULT = 'legacy',
+
+  /**
+   * card
+   */
+  CARD = 'card',
+
+  /**
+   * bubble-quote
+   */
+  BUBBLE_QUOTE = 'bubble-quote',
 }